OH MY GOD (password/email vs google)
Horrible implementation but finally I was able to differentiate whether someone didn't finish signing up with Google (with the same email) or already have a different account with the same email. The method is that if you don't finish signing up with google there's no user in the firebase database (until you sign in), whereas if you make an email/password account there's an empty (but still existent) account in firebase database.

for some reason there's a userCheckError if it doesn't exist so that's the one that gets the google error message.