Changed Item Selection and
Changed item selection to remove some silly things "Legal Documents," and the such. Also added the html and css for profile views, in the analytics section of the donation center dashboard. Plan is to have a counter incremented when it's loaded from neighbor-dashboard (new field in firebase database "profileViews" or something)